+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add `ide(type: Provider<*>, version: Provider<String>, useInstaller: Provider<Boolean>)` overload to the `pluginVerification.ides` block.
 - Warn that since the IntelliJ Platform version `2025.1` (build `251`), the required Kotlin version is `2.0.0` or higher.
+- Load the `com.intellij` bundled module by default for all IntelliJ Platform types
 
 ### Fixed