+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - Fixed the broken path resolution in the `bundledLibrary` helper and the `TestFrameworkType.Bundled` test framework
+- Fixed configuring dependencies on plugins with required content modules (#1883)
 
 ## [2.2.1] - 2025-01-21